chore: move expiry component to it's own field
