+++ frontend/components/receipts/orders/stop/submission.tsx
@@ -6,22 +6,7 @@ import { OrderTable } from '../../utils/order-table'
 import { OrderBadges } from '../../utils/order/badges'
 import { getDateTimeFormat } from '@vegaprotocol/utils'
 import { PriceWithTooltip } from '../../utils/string-amounts/price-with-tooltip'
-
-export const enum ExpiryStrategy {
-  EXPIRY_STRATEGY_UNSPECIFIED = 'EXPIRY_STRATEGY_UNSPECIFIED',
-  EXPIRY_STRATEGY_CANCELS = 'EXPIRY_STRATEGY_CANCELS',
-  EXPIRY_STRATEGY_SUBMIT = 'EXPIRY_STRATEGY_SUBMIT'
-}
-
-const EXPIRY_STRATEGY_MAP: Record<ExpiryStrategy, string> = {
-  [ExpiryStrategy.EXPIRY_STRATEGY_UNSPECIFIED]: 'Unspecified',
-  [ExpiryStrategy.EXPIRY_STRATEGY_CANCELS]: 'Cancels',
-  [ExpiryStrategy.EXPIRY_STRATEGY_SUBMIT]: 'Submit'
-}
-
-const ExpiryStrat = ({ expiryStrategy }: { expiryStrategy: ExpiryStrategy }) => {
-  return <>{EXPIRY_STRATEGY_MAP[expiryStrategy]}</>
-}
+import { ExpiryStrategyComponent } from './expiry-strategy'
 
 const SubmissionDetails = ({ title, stopOrderDetails }: { title: string; stopOrderDetails: any }) => {
   const { expiryStrategy, price, expiresAt, trailingPercentOffset, orderSubmission } = stopOrderDetails
@@ -32,7 +17,10 @@ const SubmissionDetails = ({ title, stopOrderDetails }: { title: string; stopOrd
       : null,
     trailingPercentOffset && Number(expiresAt) !== 0 ? ['Trailing offset', `${trailingPercentOffset}%`] : null,
     expiryStrategy
-      ? ['Expiry strategy', <ExpiryStrat key={`${title}-expiry-strategy`} expiryStrategy={expiryStrategy} />]
+      ? [
+          'Expiry strategy',
+          <ExpiryStrategyComponent key={`${title}-expiry-strategy`} expiryStrategy={expiryStrategy} />
+        ]
       : null,
     expiresAt && Number(expiresAt) !== 0 ? ['Expires at', getDateTimeFormat().format(new Date(+expiresAt / 1e6))] : null
   ]
